Authorization

+++ Laravel Basics/First-laravel-App/app/Http/Controllers/PostController.php
@@ -6,11 +6,23 @@ namespace App\Http\Controllers;
 // use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
+
     /**
      * Display a listing of the resource.
      */
@@ -66,6 +78,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        // Check if user has permission to edit the post
+        Gate::authorize('modify', $post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -74,6 +88,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // Only Authorized users can edit posts
+        Gate::authorize('modify', $post);
+
         // Validate
         $postFields = $request->validate([
             'title' => ['required', 'max:255'],
@@ -93,6 +110,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // Only Authorized users can delete posts
+        Gate::authorize('modify', $post);
+
         // Delete the post
         $post->delete();
